+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules/
 .expo/
 npm-debug.*
+.DS_Store
