+++ shared/components/tabs/HistoryTab.js
@@ -1,17 +1,29 @@
-import React, { PureComponent } from 'react';
-import { View, Text }           from 'react-native';
+import React, { PureComponent, PropTypes } from 'react';
+import { View, Text, TouchableWithoutFeedback } from 'react-native';
 
 import { tabBarIcon } from '../../navigation/tabsConfig.js';
 
 export default class extends PureComponent {
+    static propTypes = {
+        screenProps: PropTypes.object
+    }
+
+    handleOpenVideo = () => {
+        this.props.screenProps.mainNavigation.navigate('Video');
+    }
+
     static navigationOptions = {
         tabBarIcon: tabBarIcon.bind(this, 'History')
     }
 
     render() {
         return (
-            <View style={{ flex: 1, backgroundColor: 'yellow' }}>
-                <Text>History</Text>
+            <View style={{ flex: 1, backgroundColor: 'yellow', alignItems: 'center', justifyContent: 'center' }}>
+                <TouchableWithoutFeedback onPress={this.handleOpenVideo}>
+                    <View>
+                        <Text>Open Video</Text>
+                    </View>
+                </TouchableWithoutFeedback>
             </View>
         );
     }
